# docs/swagger.yaml
basePath: /api
definitions:
  entities.Booking:
    properties:
      computer_id:
        type: integer
      created_at:
        type: string
      end_time:
        type: string
      id:
        type: integer
      package_id:
        type: integer
      start_time:
        type: string
      status:
        type: string
      user_id:
        type: integer
    type: object
  entities.Computer:
    properties:
      cpu:
        type: string
      created_at:
        type: string
      gpu:
        type: string
      hdd:
        type: string
      headset:
        type: string
      id:
        type: integer
      keyboard:
        type: string
      monitor:
        type: string
      mouse:
        type: string
      mousepad:
        type: string
      ram:
        type: string
      ssd:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  entities.ComputerOccupiedLeftTime:
    properties:
      id:
        type: integer
      left_time:
        type: integer
    type: object
  entities.ComputerStatus:
    properties:
      id:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    type: object
  server.JSONAnswer:
    properties:
      id:
        type: integer
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for Esports Club Reservation system
  title: Esports Club Booking API
  version: "1.0"
paths:
  /api/bookings:
    get:
      consumes:
      - application/json
      description: Returns a list of bookings with various statuses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Booking'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.JSONAnswer'
      summary: Get all bookings with time format "2006-01-02T15:04:05Z07:00"
      tags:
      - Bookings
    post:
      consumes:
      - application/json
      description: Creates a new booking with the given details
      parameters:
      - description: Booking details
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/entities.Booking'
      produces:
      - application/json
      responses:
        "201":
          description: Booking created successfully
          schema:
            $ref: '#/definitions/server.JSONAnswer'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/server.JSONAnswer'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.JSONAnswer'
      summary: Create a new booking
      tags:
      - Bookings
  /api/bookings/{id}:
    get:
      consumes:
      - application/json
      description: Returns a list of bookings for the specified computer ID
      parameters:
      - description: Computer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of bookings for the computer
          schema:
            items:
              $ref: '#/definitions/entities.Booking'
            type: array
        "400":
          description: Invalid computer ID
          schema:
            $ref: '#/definitions/server.JSONAnswer'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.JSONAnswer'
      summary: Get bookings for a specific computer with time format "2006-01-02T15:04:05Z07:00"
      tags:
      - Bookings
  /api/bookings/{id}/left:
    get:
      consumes:
      - application/json
      description: Returns the left time for the specified computer ID
      parameters:
      - description: Computer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Left time for the computer
          schema:
            $ref: '#/definitions/entities.ComputerOccupiedLeftTime'
        "400":
          description: Invalid computer ID
          schema:
            $ref: '#/definitions/server.JSONAnswer'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.JSONAnswer'
      summary: Get the left time for a specific computer in seconds
      tags:
      - Bookings
  /api/bookings/active:
    get:
      consumes:
      - application/json
      description: Returns a list of active bookings
      produces:
      - application/json
      responses:
        "200":
          description: List of active bookings
          schema:
            items:
              $ref: '#/definitions/entities.Booking'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.JSONAnswer'
      summary: Get active bookings with time format "2006-01-02T15:04:05Z07:00"
      tags:
      - Bookings
  /api/bookings/cancelled:
    get:
      consumes:
      - application/json
      description: Returns a list of cancelled bookings
      produces:
      - application/json
      responses:
        "200":
          description: List of cancelled bookings
          schema:
            items:
              $ref: '#/definitions/entities.Booking'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.JSONAnswer'
      summary: Get cancelled bookings with time format "2006-01-02T15:04:05Z07:00"
      tags:
      - Bookings
  /api/bookings/computers-left-time:
    get:
      consumes:
      - application/json
      description: Returns a list of computers left time
      produces:
      - application/json
      responses:
        "200":
          description: List of computers left time
          schema:
            items:
              $ref: '#/definitions/entities.ComputerOccupiedLeftTime'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.JSONAnswer'
      summary: Get computers left time in seconds
      tags:
      - Bookings
  /api/bookings/finished:
    get:
      consumes:
      - application/json
      description: Returns a list of finished bookings
      produces:
      - application/json
      responses:
        "200":
          description: List of finished bookings
          schema:
            items:
              $ref: '#/definitions/entities.Booking'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.JSONAnswer'
      summary: Get finished bookings with time format "2006-01-02T15:04:05Z07:00"
      tags:
      - Bookings
  /api/bookings/pending:
    get:
      consumes:
      - application/json
      description: Returns a list of pending bookings
      produces:
      - application/json
      responses:
        "200":
          description: List of pending bookings
          schema:
            items:
              $ref: '#/definitions/entities.Booking'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.JSONAnswer'
      summary: Get pending bookings with time format "2006-01-02T15:04:05Z07:00"
      tags:
      - Bookings
  /api/computers:
    get:
      description: Returns a list of all computers
      produces:
      - application/json
      responses:
        "200":
          description: List of computers
          schema:
            items:
              $ref: '#/definitions/entities.Computer'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.JSONAnswer'
      summary: Get all computers
      tags:
      - Computers
    post:
      consumes:
      - application/json
      description: Creates a new computer
      parameters:
      - description: Computer details
        in: body
        name: computer
        required: true
        schema:
          $ref: '#/definitions/entities.Computer'
      produces:
      - application/json
      responses:
        "201":
          description: Computer created successfully
          schema:
            $ref: '#/definitions/server.JSONAnswer'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/server.JSONAnswer'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.JSONAnswer'
      summary: Create a new computer
      tags:
      - Computers
  /api/computers/{id}:
    delete:
      description: Deletes a specific computer by ID
      parameters:
      - description: Computer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Computer deleted successfully
          schema:
            $ref: '#/definitions/server.JSONAnswer'
        "400":
          description: Invalid computer ID
          schema:
            $ref: '#/definitions/server.JSONAnswer'
        "404":
          description: Computer not found
          schema:
            $ref: '#/definitions/server.JSONAnswer'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.JSONAnswer'
      summary: Delete a specific computer
      tags:
      - Computers
    get:
      description: Returns a specific computer by ID
      parameters:
      - description: Computer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Computer details
          schema:
            $ref: '#/definitions/entities.Computer'
        "400":
          description: Invalid computer ID
          schema:
            $ref: '#/definitions/server.JSONAnswer'
        "404":
          description: Computer not found
          schema:
            $ref: '#/definitions/server.JSONAnswer'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.JSONAnswer'
      summary: Get a specific computer
      tags:
      - Computers
  /api/computers/{id}/status:
    get:
      consumes:
      - application/json
      description: Returns the status of the specified computer ID
      parameters:
      - description: Computer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Computer status
          schema:
            $ref: '#/definitions/entities.ComputerStatus'
        "400":
          description: Invalid computer ID
          schema:
            $ref: '#/definitions/server.JSONAnswer'
        "404":
          description: Computer not found
          schema:
            $ref: '#/definitions/server.JSONAnswer'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.JSONAnswer'
      summary: Get the status of a specific computer
      tags:
      - Computers
  /api/computers/statuses:
    get:
      consumes:
      - application/json
      description: Returns a list of all computer statuses
      produces:
      - application/json
      responses:
        "200":
          description: List of computer statuses
          schema:
            items:
              $ref: '#/definitions/entities.ComputerStatus'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.JSONAnswer'
      summary: Get all computer statuses
      tags:
      - Computers
swagger: "2.0"